Handle errors explicitly
First, this change updates database stats recording to return early if
there is an error when creating instruments. Before, the resulting
instrument was nil and RegisterCallback returned an error instead.

Second, it adds oteltest.NewMetricProviderWithError to help with testing
this behavior.

It also documents panic behavior for Wrap function and removes
non-idiomatic handleErr functions.